test_project/requirements.txt
attrs==21.4.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
tomli==2.0.1
pytest-xdist
selenium==4.1.5
webdriver-manager==3.5.4
requests~=2.27.1
allure-pytest
jsonschema
faker==13.11.1
SQLAlchemy~=1.4.36
pymysql
Flask~=2.1.2